split out components into separate modules, fix file names
