--- src/innov8/run.py
@@ -1,1 +1,0 @@
-# Run the app
